+++ lib/src/obsoleted/rules/avoid_late_keyword.dart
@@ -1,0 +1,60 @@
+import 'package:analyzer/dart/analysis/results.dart';
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
+
+import '../../models/issue.dart';
+import '../../models/severity.dart';
+import '../../utils/node_utils.dart';
+import '../../utils/rule_utils.dart';
+import 'obsolete_rule.dart';
+
+class AvoidLateKeywordRule extends ObsoleteRule {
+  static const String ruleId = 'avoid-late-keyword';
+  static const _documentationUrl = '';
+
+  static const _warning = "Avoid using 'late' keyword.";
+
+  AvoidLateKeywordRule({Map<String, Object> config = const {}})
+      : super(
+          id: ruleId,
+          documentationUrl: Uri.parse(_documentationUrl),
+          severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
+        );
+
+  @override
+  Iterable<Issue> check(ResolvedUnitResult source) {
+    final visitor = _Visitor();
+
+    source.unit?.visitChildren(visitor);
+
+    return visitor.declarations
+        .map((declaration) => createIssue(
+              rule: this,
+              location: nodeLocation(
+                node: declaration,
+                source: source,
+                withCommentOrMetadata: true,
+              ),
+              message: _warning,
+            ))
+        .toList(growable: false);
+  }
+}
+
+class _Visitor extends RecursiveAstVisitor<void> {
+  final _declarations = <AstNode>[];
+
+  Iterable<AstNode> get declarations => _declarations;
+
+  @override
+  void visitVariableDeclaration(VariableDeclaration node) {
+    super.visitVariableDeclaration(node);
+
+    if (node.isLate && node.parent != null) {
+      final parent = node.parent;
+
+      _declarations.add(parent ?? node);
+    }
+  }
+}